chore: abstract ConfigScreen's form
